Fix play svp video display picture
The OverView status video should continue to play after recent repeated,switches from Overview to Allapps status. When RecentsView was reset,it switched to the image display mode. Causing the launcher to overwrite the video.

The scheme is to deternmine that if it is a three-key navigation mode and the animation is cancel,setCurrentTask(-1) will not be executed.

Bug: 298512495
Test: manual

Change-Id: I6cc419c011144e9ce6b0f4ab6d9ac802a174db0e